requirements.txt
Arrays==0.0.5
asyncio==3.4.3
qtm==2.1.1
sockets==1.0.0
minimalmodbus~=2.0.1
wheel==0.40.0
future~=0.18.3
PyYAML~=6.0
pyserial~=3.5
iso8601~=1.1.0
libscrc~=1.8.1
numpy~=1.24.3
scipy~=1.10.1
matplotlib
jupyterlab
notebook
jupyter
